+++ budget-binder-multiplatform-app/src/jvmMain/kotlin/de/hsfl/budgetBinder/Router.kt
@@ -20,7 +20,7 @@ fun Router() {
         is Screen.Register -> RegisterComponent()
         is Screen.Login -> LoginComponent()
         is Screen.Dashboard -> DashboardComponent()
-        is Screen.Settings.Menu, Screen.Settings.User -> SettingsView()
+        is Screen.Settings.Menu, Screen.Settings.User, Screen.Settings.Server -> SettingsView()
         else -> {}
     }
 }
