+++ budget-binder-multiplatform-app/src/jvmMain/kotlin/de/hsfl/budgetBinder/compose/textfield/SettingsTextField.kt
@@ -4,45 +4,44 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.Icon
 import androidx.compose.material.MaterialTheme
-import androidx.compose.material.OutlinedTextField
 import androidx.compose.material.Text
+import androidx.compose.material.TextField
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Info
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.input.VisualTransformation
 
 @Composable
-fun EmailTextField(
+fun SettingsTextField(
     modifier: Modifier = Modifier,
     value: String,
     onValueChange: (String) -> Unit,
-    label: @Composable (() -> Unit)? = null,
+    label: @Composable () -> Unit,
+    isError: Boolean = false,
+    errorText: String = "",
     enabled: Boolean = true,
-    isError: Boolean = false
+    visualTransformation: VisualTransformation = VisualTransformation.None,
+    keyboardOptions: KeyboardOptions = KeyboardOptions.Default,
 ) {
     Column {
-        OutlinedTextField(
+        TextField(
             modifier = modifier,
             value = value,
             onValueChange = onValueChange,
-            enabled = enabled,
-            isError = isError,
             label = label,
-            singleLine = true,
+            isError = isError,
+            enabled = enabled,
+            visualTransformation = visualTransformation,
+            keyboardOptions = keyboardOptions,
             trailingIcon = {
                 if (isError) {
                     Icon(Icons.Default.Info, contentDescription = null)
                 }
-            },
-            keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Email)
+            }
         )
-    }
-    if (isError) {
-        Text(
-            text = "Please enter a valide email address",
-            color = MaterialTheme.colors.error,
-            style = MaterialTheme.typography.caption
-        )
+        if (isError) {
+            Text(text = errorText, style = MaterialTheme.typography.caption, color = MaterialTheme.colors.error)
+        }
     }
 }
